requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# HTTP client
httpx==0.27.2

# Config
python-dotenv==1.0.1

# Auth
passlib[bcrypt]==1.7.4

# Frontend
streamlit==1.40.0
pandas==2.2.3
plotly==5.24.1

# AI / ML
sentence-transformers==3.3.1
chromadb==0.5.20

# Invoice generation
weasyprint==62.3

# Caching
redis==5.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
